+++ src/ConductorSharp.Engine/Builders/Metadata/OriginalNameAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ConductorSharp.Engine.Util
+namespace ConductorSharp.Engine.Builders.Metadata
 {
     [AttributeUsage(AttributeTargets.Class)]
     public class OriginalNameAttribute : Attribute
