+++ src/ConductorSharp.Engine/Builders/Metadata/VersionAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace ConductorSharp.Engine.Builders.Metadata;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class VersionAttribute(int version) : Attribute
+{
+    internal int Version { get; } = version;
+}
